Optimise getTreeIntervals for big trees.
New implementation results in at least 2x speedup for tree of 15K tips.
Speedup is mostly from initialising in ways that can use primitives.
The new implementation also uses an adjacency list instead of searching
the treetable for a node's children, but this leads to a relatively minor
speedup.

Further speedups may be possible by restricting the use case to a binary
tree (only 2 or no children), or by making the treetable a global var and
using lapply (use primitives) or mclapply (use multiple cores).